+++ AteliApp.Application/DTOs/AtualizarMedidasDto.cs
@@ -2,20 +2,20 @@
 
 public class AtualizarMedidasDto
 {
-    public int? Altura { get; }
-    public int? Busto { get; }
-    public int? Cintura { get; }
-    public int? Quadril { get; }
-    public int? Ombro { get; }
+    public decimal? Altura { get; }
+    public decimal? Busto { get; }
+    public decimal? Cintura { get; }
+    public decimal? Quadril { get; }
+    public decimal? Ombro { get; }
     public string? Outros { get; }
     public string? Observacao { get; }
 
     public AtualizarMedidasDto(
-        int? altura = null,
-        int? busto = null,
-        int? cintura = null,
-        int? quadril = null,
-        int? ombro = null,
+        decimal? altura = null,
+        decimal? busto = null,
+        decimal? cintura = null,
+        decimal? quadril = null,
+        decimal? ombro = null,
         string? outros = null,
         string? observacao = null)
     {
